plan2walls/requirements.txt
ruamel-yaml
accelerate
bitsandbytes
datasets
geonamescache
git+https://github.com/facebookresearch/segment-anything.git
googletrans==3.1.0a0
google-cloud-storage
google-auth
google-cloud-vision
img2dataset
linktransformer
matplotlib
onnx
onnxruntime
opencv-python
pandas
Pillow
plotly==5.18.0
pycocotools
pycountry==20.7.3
pytorch-lightning
seaborn
sentence-transformers
seqeval
simplejson
tensorboard
timm
torch==2.6.0
torchvision==0.21.0
transformers
scikit-learn==1.7.2
tokenizers==0.22.1
fastapi
uvicorn
diffusers
python-multipart
# --upgrade diffusers[torch]
# -U kaleido